Make it clear from arguments that attitude is required
An attitude is required for the star_plot routine to plot stars, and
the routine now always has stars (either supplied or fetched).
Theoretically, one could avoid using the attitude if one has already
determined yag/zag positions for the stars in the calling code,
but the attitude should be known, so this isn't a burden on the user.